Rename repository "craniodistractor" as "cranio" (#9)
