+++ run.py
@@ -2,9 +2,9 @@ import sys
 import random
 import datetime
 
-from craniodistractor.producer import ProducerProcess
-from craniodistractor.imada import ImadaSensor
-from craniodistractor.app.plot import PlotWindow, update_plot, time_filter, PlotWidget
+from cranio.producer import ProducerProcess
+from cranio.imada import ImadaSensor
+from cranio.app.plot import PlotWindow, update_plot, time_filter, PlotWidget
 
 start_time = datetime.datetime.now()
 n_seconds = 3
@@ -13,7 +13,7 @@ def update(plot_widget):
     time_filter(n_seconds, update_plot(plot_widget, x=[(datetime.datetime.now()-start_time).total_seconds()], y=[random.gauss(0,1)]))
 
 def run():
-    ''' Runs the craniodistractor prototype '''
+    ''' Runs the cranio prototype '''
     p = PlotWindow()
     p.ok_button.setText('Analyze')
     w = PlotWidget()
