+++ setup.py
@@ -1,8 +1,8 @@
 from setuptools import setup
 
 setup(
-    name='craniodistractor',
-    packages=['craniodistractor'],
+    name='cranio',
+    packages=['cranio'],
     install_requires=['pandas', 'pyqtgraph', 'pyserial'],
     entry_points={
         'console_scripts': [],
