+++ test/test_imada.py
@@ -1,4 +1,4 @@
-from craniodistractor.imada import decode_telegram
+from cranio.imada import decode_telegram
 
 def test_decode_telegram():
     assert (-1.234, 'K', 'T', 'O') == decode_telegram('-1.234KTO\r')
